big a probable bug with backups v2

+++ plogical/IncScheduler.py
@@ -1501,7 +1501,8 @@ Automatic Backupv2 failed for %s on %s.
                                     else:
                                         value['lastRun'] = time.strftime("%m.%d.%Y_%H-%M-%S")
 
-                                    background.DeleteSnapshots(f"--keep-daily {value['retention']}")
+                                    if function == '1 Week':
+                                        background.DeleteSnapshots(f"--keep-daily {value['retention']}")
                             except BaseException as msg:
                                 print("Error: [v2Backups]: %s" % str(msg))
                                 logging.writeToFile('%s. [v2Backups]' % (str(msg)))
